* Refactor ssl_connected() with comments and clarity. * Fetch a SSL server's certificate (including subject, issuer, pkey, and digest) * Extend /on ssl_server_cert with $6=SSL digest (+URLified)) * Someday, we'll save the ssl_certificate and cert_hash in the server struct * Someday, we want ssl.h to save all the metadata, for non-servers. * Add new argument "snap" to add_timer(), for determining initial timeout. * Updated tmux_away, userlist script (zlonix) * Fix bug in target_file_write() not properly handling @W/@L/@ref * Refactor the timer subsystem with comments and clarity * Add /TIMER -SNAP which allows you to run timers "at the top of the interval" * Make /timer option processing similar to other commands * Add comments to make the timer API more clear and robust. * Revamp add_time() to add snapping, and handle updating better * Add /SET SSL_ROOT_CERT_FILE to pass to SSL_CTX_load_verify_locations * Add new script 'find_ssl_root_certs' which sets SSL_ROOT_CERT_FILE for you. * Have 'find_ssl_root_certs' loaded by 'global' * Fix very many warnings recommended by clang.

+++ include/ircaux.h
@@ -7,7 +7,7 @@
  *
  * See the COPYRIGHT file, or do a HELP IRCII COPYRIGHT 
  *
- * @(#)$Id: ircaux.h,v 1.129 2014/04/24 04:51:42 jnelson Exp $
+ * @(#)$Id: ircaux.h,v 1.130 2015/04/11 04:16:33 jnelson Exp $
  */
 
 #ifndef _IRCAUX_H_
@@ -316,7 +316,7 @@ extern	int     quick_code_point_count	(const unsigned char *str);
 extern	int     previous_code_point	(const unsigned char *, const unsigned char **);
 extern	int	quick_code_point_index	(const unsigned char *, const unsigned char *);
 
-extern	int	strext2		(unsigned char **, unsigned char *, int , int);
+extern	int	strext2		(unsigned char **, unsigned char *, size_t , size_t);
 extern	int     invalid_utf8str (unsigned char *utf8str);
 extern	int     is_iso2022_jp (const unsigned char *buffer);
 extern	void    create_utf8_locale (void);
